Reorganized backend tests, starting front end tests
